# Replace some unsafe with safe Rust without performance regression