--- a/src/lib.rs
+++ b/src/lib.rs
@@ -132,16 +132,7 @@ fn setout(src: &[u8], dst: &mut [u8], len: usize) {
 
 fn xorin(dst: &mut [u8], src: &[u8]) {
     assert!(dst.len() <= src.len());
-    let len = dst.len();
-    let mut dst_ptr = dst.as_mut_ptr();
-    let mut src_ptr = src.as_ptr();
-    for _ in 0..len {
-        unsafe {
-            *dst_ptr ^= *src_ptr;
-            src_ptr = src_ptr.offset(1);
-            dst_ptr = dst_ptr.offset(1);
-        }
-    }
+    dst.iter_mut().zip(src.iter()).for_each(|(a, b)| *a ^= b);
 }
 
 /// Total number of lanes.