There seems to be no performance regression. I actually noticed a small improvement.

With unsafe:
```
test bench_sha3_256_input_4096_bytes ... bench:      13,141 ns/iter (+/- 890) = 311 MB/s
test keccakf_u64                     ... bench:         413 ns/iter (+/- 31) = 60 MB/s
```

Without unsafe:
```
test bench_sha3_256_input_4096_bytes ... bench:      12,986 ns/iter (+/- 1,054) = 315 MB/s
test keccakf_u64                     ... bench:         409 ns/iter (+/- 27) = 61 MB/s
```