@debris 2018-11-26
I'll double check and merge if it is indeed faster. With an old version of compiler (before rust 1.20) it was muuuch slower