@brycx 2018-11-26
Sounds great!

Also, do you have any news about the big-endian fix (issue: #15 , possible fix: #35)? I'm using tiny-keccak in another library myself and would like to know if you plan on resolving the issue.