@burdges 2018-12-20
Yes, these unsafe loops do not elide bounds checks correctly.  I submitted a fix in https://github.com/debris/tiny-keccak/pull/8/commits/ccf709b13ed1e9512068c624e707cdc7d971ed66 that was idiomatic for highly optimized Rust, but I'm happy the compiler has improved its optimizations of truely idiomatic Rust.